# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Skill Radar API
  description: |
    Findet vorhandene Kompetenzen (Mitarbeiter/Skills) und dokumentiertes Wissen
    (Artikel/Anhänge) im Unternehmen auffindbar. Begriffe folgen dem Glossar in
    `CONTEXT.md` (Mitarbeiter, Skill, Selbsteinschätzung, Artikel, Mitwirkender, Anhang).

    This document specifies the implemented API derived from the repo's GitHub
    issues (`OfficerYoda/basf-hackathon`).

    All endpoints exchange JSON (`application/json`) and report errors as
    `{"error": "human-readable message"}`, except file upload/download on
    attachments.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development (docker compose + go run ./backend)
security: []
...
paths:
...
  /api/employees:
    get:
      tags: [Employees]
      summary: List employees
      description: "✅ Implemented. Returns all employees ordered by id, each with its skills."
      operationId: listEmployees
      responses:
        "200":
          description: List of employees.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      tags: [Employees]
      summary: Create an employee
      description: |
        ✅ Implemented. Request bodies are decoded strictly: unknown fields and any
        trailing content after the JSON object are rejected with 400. Body is
        capped at 1 MiB. Skill names are lowercased before storage; duplicate skill
        names (post-lowercasing) within the same request are rejected with 400.
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      responses:
        "201":
          description: Employee created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "email or skill already exists"
  /api/employees/{id}:
    parameters:
      - $ref: "#/components/parameters/EmployeeId"
    get:
      tags: [Employees]
      summary: Get an employee
      description: "✅ Implemented."
      operationId: getEmployee
      responses:
        "200":
          description: The employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Employees]
      summary: Replace an employee
      description: |
        ✅ Implemented. Full replace: the submitted `skills` array replaces the
        employee's existing skills entirely (skills not included are removed).
      operationId: updateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      responses:
        "200":
          description: Updated employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Email already in use by another employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Employees]
      summary: Delete an employee
      description: |
        ✅ Implemented. Cascades to the employee's skills. Rejected if the employee
        is a contributor on any article — remove them from the article(s) first
        (issue #2's article_contributors has `ON DELETE RESTRICT`).
      operationId: deleteEmployee
      responses:
        "204":
          description: Employee deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Employee is a contributor on at least one article.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "employee is referenced by an article"
  /api/articles:
    get:
      tags: [Articles]
      summary: List articles
      description: "✅ Implemented (issues #2 and #3). Returns all articles ordered by id, each with its skills, contributors, and attachments."
      operationId: listArticles
      responses:
        "200":
          description: List of articles, each with its skills, contributors, and attachments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      tags: [Articles]
      summary: Publish an article
      description: |
        ✅ Implemented (issue #2). `skills` must reference skill names already
        present in the shared skill vocabulary (i.e. used by at least one
        employee); `contributor_ids` must reference existing employees. Authors
        and other participants are not distinguished by role — all are
        "Mitwirkende". Request bodies are decoded strictly, same as employees
        (unknown fields and trailing content rejected with 400, 1 MiB cap).
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Article created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Invalid body, or a referenced skill/contributor does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/articles/{id}:
    parameters:
      - $ref: "#/components/parameters/ArticleId"
    get:
      tags: [Articles]
      summary: Get an article
      description: "✅ Implemented (issues #2 and #3). Includes skills, contributors, and attachments."
      operationId: getArticle
      responses:
        "200":
          description: The article.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Articles]
      summary: Replace an article
      description: "✅ Implemented (issue #2). Full replace of title, description, content, skills, and contributors."
      operationId: updateArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "200":
          description: Updated article.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Invalid body, or a referenced skill/contributor does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Articles]
      summary: Delete an article
      description: |
        ✅ Implemented (issues #2 and #3). Cascades to the article's skill tags,
        contributor links, attachment metadata, and stored attachment files.
      operationId: deleteArticle
      responses:
        "204":
          description: Article deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/articles/{articleId}/attachments:
    parameters:
      - $ref: "#/components/parameters/ArticleIdForAttachment"
    post:
      tags: [Attachments]
      summary: Upload an attachment
      description: |
        ✅ Implemented (issue #3). Stores the file on a local volume and its
        metadata in Postgres, linked to the article. No
        maximum upload size is enforced, per the product decision for this
        hackathon MVP. The stored filename is derived server-side so a client
        cannot escape the storage directory or overwrite an existing file.
      operationId: createAttachment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to attach. Original filename and MIME type are read from the multipart part.
      responses:
        "201":
          description: Attachment stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "400":
          description: Missing file part or invalid multipart body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/attachments/{id}:
    parameters:
      - $ref: "#/components/parameters/AttachmentId"
    get:
      tags: [Attachments]
      summary: Download an attachment
      description: |
        ✅ Implemented (issue #3). Streams the stored file. The response `Content-Type`
        reflects the attachment's stored MIME type and `Content-Disposition`
        carries the original filename.
      operationId: downloadAttachment
      responses:
        "200":
          description: File contents.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Attachment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Attachments]
      summary: Delete an attachment
      description: "✅ Implemented (issue #3). Removes both the stored file and its metadata; never leaves one without the other."
      operationId: deleteAttachment
      responses:
        "204":
          description: Attachment deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Attachment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search:
    get:
      tags: [Search]
      summary: Search employees and articles
      description: |
        ✅ Implemented (issue #4). Free text (`q`) matches employee names, and article
        titles or markdown content. Skill filters (`skills`) narrow the result to
        entries holding all (`mode=and`) or at least one (`mode=or`) of the given
        skills; skill names are matched case-insensitively against their
        lowercase-normalized form. `q` and `skills` may be combined: a result must
        satisfy the text match (if `q` is given) *and* the skill filter (if
        `skills` is given).

        Deterministic behavior for edge cases: omitting both `q` and `skills`
        returns all employees and articles.
        An empty `skills` list is treated as "no skill filter".
      operationId: search
      parameters:
        - name: q
          in: query
          required: false
          description: Free-text query, matched against employee names and article titles/content.
          schema:
            type: string
        - name: skills
          in: query
          required: false
          description: Repeat this parameter for each skill name to filter by.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          example: [python, leadership]
        - name: mode
          in: query
          required: false
          description: How multiple `skills` combine.
          schema:
            type: string
            enum: [and, or]
            default: and
      responses:
        "200":
          description: Matches, grouped by type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Invalid `mode` value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/skills/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: Skill name. Lowercased and trimmed server-side before matching.
        schema:
          type: string
    put:
      tags: [Skills]
      summary: Register a skill name
      description: |
        ✅ Implemented. Idempotent, no request body: ensures the skill exists in
        the shared vocabulary so an article can reference it via `skills` in
        `ArticleInput` even before any employee has self-assessed it. Calling
        this again for an already-registered name is a no-op and still returns
        200.
      operationId: putSkill
      responses:
        "200":
          description: The skill now exists (created or already present).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillName"
        "400":
          description: Blank skill name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Skills]
      summary: Remove a skill name
      description: |
        ✅ Implemented. Rejected with 409 while any employee or article still
        references the skill (foreign key constraint) — remove those
        references first.
      operationId: deleteSkill
      responses:
        "204":
          description: Skill deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Skill is still used by an employee or article.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "skill is used by an employee or article"
components:
  parameters:
    EmployeeId:
      name: id
      in: path
      required: true
      description: Employee id.
      schema:
        type: integer
        format: int64
        minimum: 1
    ArticleId:
      name: id
      in: path
      required: true
      description: Article id.
      schema:
        type: integer
        format: int64
        minimum: 1
    ArticleIdForAttachment:
      name: articleId
      in: path
      required: true
      description: Id of the article the attachment belongs to.
      schema:
        type: integer
        format: int64
        minimum: 1
    AttachmentId:
      name: id
      in: path
      required: true
      description: Attachment id.
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    BadRequest:
      description: Invalid request body or parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: employee not found
...
    Employee:
      type: object
      required: [name, email]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          minLength: 1
          example: Ada Lovelace
        email:
          type: string
          format: email
          example: ada@example.com
        department:
          type: string
          example: Digitalization
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
          default: []
          description: Skill names must be unique within the employee (case-insensitive, after lowercasing).
    ArticleInput:
      type: object
      required: [title, description, content]
      properties:
        title:
          type: string
          minLength: 1
          example: Onboarding auf den neuen Deploy-Prozess
        description:
          type: string
          minLength: 1
          example: Kurzüberblick über den neuen CI/CD-Ablauf.
        content:
          type: string
          minLength: 1
          description: Markdown body.
          example: "## Übersicht\n\n..."
        skills:
          type: array
          items:
            type: string
          default: []
          description: >
            Names of skills already present in the shared skill vocabulary (i.e.
            used by at least one employee). Lowercased before matching; an unknown
            name is rejected with 400.
          example: [ci-cd, deployment]
        contributor_ids:
          type: array
          items:
            type: integer
            format: int64
          default: []
          description: >
            Ids of employees credited on this article. Authors and other
            participants are not distinguished by role (all are Mitwirkende).
    Article:
      type: object
      required: [id, title, description, content, skills, contributors, attachments, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
        description:
          type: string
        content:
          type: string
        skills:
          type: array
          items:
            type: string
        contributors:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
          description: >
            Full employee records (including their own skills) for each
            contributor, not just a name/id reference.
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
          description: Attachments ordered by id.
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    SearchResponse:
      type: object
      required: [employees, articles]
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
    SkillName:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Lowercased, trimmed skill name.
          example: rust